fix(web): test parseRangeHeader without importing the auth and DB stack
The unit job failed on PR #61 with "DATABASE_URL is not set", thrown from
database/connection.ts at module load.

range-header.test.ts imported parseRangeHeader from storage/fileStorage, which
imports auth/server, which imports database/connection — and that throws
immediately when DATABASE_URL is unset. So a test for a regex and some arithmetic
was standing up the auth stack and a PG pool.

It passed locally and only ever failed in CI because web-app/code/.env exists on a
dev machine and dotenvx injects it; the CI unit job has no .env by design — it is
documented as "Unit tests — web (jsdom) + mobile (node). No database needed."

Moved the function to storage/range.ts, which imports only ./provider (types,
zero imports). fileStorage imports it from there; the one production call site is
unchanged. The CI invariant is now structurally true rather than accidentally
maintained: nothing about testing this function can pull in a database again.

Reproduced the failure locally by hiding .env, and verified the full web unit
suite passes the same way: 106 passed, previously 91 passed with one suite failing
to collect.
